+++ packages/bpk-component-radio/stories.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { storiesOf, action } from '@kadira/storybook';
+import { colorGray700, spacingBase } from 'bpk-tokens/tokens/base.es6';
 
 import BpkRadio from './index';
 
@@ -20,6 +21,29 @@ storiesOf('bpk-component-radio', module)
       label="Return"
       onChange={action('radio changed')}
     />
+  ))
+  .add('White (Checked)', () => (
+    <div style={{ backgroundColor: colorGray700, padding: spacingBase }}>
+      <BpkRadio
+        id="checked"
+        name="checked"
+        label="Return"
+        onChange={action('radio changed')}
+        white
+        checked
+      />
+    </div>
+  ))
+  .add('White (Unchecked)', () => (
+    <div style={{ backgroundColor: colorGray700, padding: spacingBase }}>
+      <BpkRadio
+        id="unchecked"
+        name="unchecked"
+        label="Return"
+        onChange={action('radio changed')}
+        white
+      />
+    </div>
   ))
   .add('Disabled (Checked)', () => (
     <BpkRadio